+++ src/screens/game_screen/square.rs
@@ -3,17 +3,17 @@ use ggez::graphics::Color;
 
 #[derive(Debug, Clone, Copy)]
 pub struct Square {
-    pub row: u16,
-    pub column: u16,
+    pub row: u8,
+    pub column: u8,
     pub color: Color,
 }
 
 impl Square {
-    pub fn new(row: u16, column: u16, color: Color) -> Square {
+    pub fn new(row: u8, column: u8, color: Color) -> Square {
         Square { row, column, color }
     }
 
-    pub fn default(row: u16, column: u16) -> Self {
+    pub fn default(row: u8, column: u8) -> Self {
         Square {
             row,
             column,
